+++ ai4j-spring-boot-starter/src/main/java/io/github/lnyocly/ai4j/AiPlatformProperties.java
@@ -26,4 +26,5 @@ public class AiPlatformProperties {
     private String musicUrl;
     private String lyricsUrl;
     private String fetchUrl;
+    private String systemOneUrl;
 }